requirements/local.txt
asgiref==3.11.1
attrs==25.4.0
Django==5.2.11
django-environ==0.12.1
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pluggy==1.6.0
psycopg==3.3.2
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
